This patch is a re-work of patch r1538. One one hand, the plain old behaviour of setting the session size in ui_init is not going to work, due to xrandr, as pointed out in r1538. However, the approach implemented in this revision doesn't work either: When the window is created before the connection, this means that as soon as X11 events are recieved, this is going to trigger RDP transmissions. For example, a call to reset_modifiers_keys. But if the RDP connection is not ready, the WTS is not prepared to handle such data. We must wait with, for example, keyboard input until the connection is READY. OTOH, we can't just ignore those X11 events; that might lead to that we are not sending information that we need to send.
So, it is actually better to wait with creating the window until we
have been connected. An additional advantage of this is that for the
load balancing / session directory case, there's no risk of confusion
of which RDP connection we are actually sending data to. 

The previous behaviour of creating the window after we have been
connected has been restored.

Since we still need to set connection data (currently screen size) on
a per connection basis, we need to create a new UI function for this.

Non-X11 backends need to implement this new function.