Remove process of packets in rdp_connect.
The process of packets in rdp_connect assumes that the
first packets are demand active to set g_rdp_shareid which
is not true. In case when CredSSP+Kerberos is in use an
enhanced redirection PDU packet is sent before the demande
active PDU.

This fixes the problem with rdesktop silently exists in
the case above.